feat(utility): useIsMobile function for easy&quick implementation of responsive design
